% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{require_backend}
\alias{require_backend}
\alias{require_gpu}
\alias{require_cuda}
\alias{require_metal}
\alias{require_rocm}
\title{Require a GPU or vendor backend}
\usage{
require_gpu()

require_cuda()

require_metal()

require_rocm()
}
\value{
`TRUE` invisibly, or an error.
}
\description{
Require a GPU or vendor backend
}
