% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{visible_devices}
\alias{visible_devices}
\title{Report accelerator visibility controls}
\usage{
visible_devices()
}
\value{
A named character vector. Unset variables are represented by `NA`.
}
\description{
Report accelerator visibility controls
}
